Move main.cpp into project Root
Given all the non-main files are setup to be imported as a library in CMake anyway, it seems a little wrong for the project main to be in that same directory. It also means the built executable will no longer be put in a 'src' subdirectory of the build directory!