Create Console.h/cpp to handle ANSI codes
ANSI support is nice, especially as it's one of the easier ways to implement sensible arrow-key handling, and comes with a whole suite of extra QoL features - fonts, colours etc.

Unfortunately, while there is some support for them in Windows Console, they need to be specifically enabled with windows specific SetConsoleMode calls, and then written to with the windows WriteConsole function.

Likewise, setting a console to 'raw' mode (to be able to process such ANSI codes without waiting for an end-line) is done differently in Windows from posix terminals.

I looked into doing this using Ncurses, or PDcurses, but wasn't able to get them to work for both Windows Console and Mintty.

So, to handle this in a somewhat sane manner, I've implemented a Console module to deal with that complication, and added a -DWINBUILD option to the CMakeLists, to tell the compiler to build for windows. It includes a printCP() function to allow everything else to not worry about the underlying console.