Fixed 'raw' mode on Windows
Found a working system to put the terminal into 'raw' mode on Windows, by bracketing the termios method with  code to save and restore the Windows ConsoleMode (as indicated by the MinimumRawMode methods in main.cpp).

Updated Console.cpp to use this method. I don't particularly like it, but it works consistently, and means I don't deal with an enormous table of Windows keycodes.